Consolidate markets into constants
